Create Codes and input files
This folder contains the sources of the three programs needed to compute the spectrum of rapidly rotating stars with radiatives envelopes, and the corresponding input files. 

star3d: parameterises the surface of the star, computes radius, temperature and effective gravity as a function of the colatitude

spec3d: carries out the computation of the emergent spectrum

normspec: normalises the spectrum placing the continuum at intensity 1.0

See file README_inputs_outputs.txt in this folder for further details.